requirements.txt
asgiref==3.5.0
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5